Refactor CMake code and make it more local
"More local" means that the CMake code is spread out over multiple
CMakeLists (CML) and the code is right next to where the "information"
is. E.g. sources are added to targets in the CML file that is in the
same directory as the source files themselves and not in the top level
file.