Completely rework persistent state
Every variable has its own specific getter and setter function now. It
is also no longer a class because it doesn't need to be. Due to Rodos
enforcing the use of global variables that must be initialized in some
thread's init() the advantage of classes which is establishing an
invariant during construction is no longer there.