Move subscribers to same file as topics
This is important to prevent the static initialization order fiasco. The
subscribers' initialization depends on the topics so it must be
guaranteed that the topics are initialized before the subscribers. This
is only possible if both global variables are in the same translation
unit (file) because only there is the order of initialization well
defined (top to bottom). In different translation units the order is
ambiguous.