Allow switching between empty and looped EDU queue
The type of EDU program queue is chosen by selecting which source file
to be compiled, either the LoopedEduProgramQueue.cpp or the
EmptyEduProgramQueue.cpp. The golden test uses the empty queue since
that yields consitent expected output and for MFTs or HAFs the looped
EDU program queue can be used and easily adapted without making the
automated tests fail.