Always specify a dummy NSS pin to fix RF-EDU issue
The RF-EDU issue is acutally an issue between all SPIs and UART1. When
an SPI is initialized the default value for the NSS pin (GPIO_INVALID)
is not handled correctly and sets PA15 to the alternative function AF05.
This fucks up the configuration for UART1 where we use PA15 as TX pin.