Create new folder/component FramSections
The Periphery component was getting too full and contained too many
levels of abstraction. I decided to keep only the low-level drivers for
the peripherals in the Periphery component and move the new,
higher-level abstractions for the FRAM sections to a new component
called FramSections. PersistentVariableInfo<> is actually on an even
higher level but I'll keep it in FramSections for now. When the full
implementation of the persistent variables as well as the EDU and
telemetry memories are done I'll decide if I want to split this up
further.