Fix `RingArray::FindAndReplace()` calling public functions
Since `RingArray` uses the Thread-safe Interface pattern, public
functions must not call other public functions because that would lead
to recursive locking. Rodos' semaphores might support that but I think
it is cleaner and easier to reason about if we stick to the pattern and
ensure that we only ever enter the semaphore once.

`FindAndReplace()` used `Get()` and `Set()` so I had to make private
implementation functions (`DoGet()` and `DoSet()`) for them that do all
the work except for creating the `ScopeProtector`.