Remove `flash::Enter4ByteAdressMode()`
We use the explicit 4-byte address commands anyway so it makes no
difference which mode we are in. Interestingly, status register 3 now
reads 0x41 instead 0x40. According to the datasheet this would mean that
we are now in 4-byte mode and previously were in 3-byte mode. This is
weird and unexpected but errors in datasheets are not unheard of.