Fix `CMakeDeveloperPresets.json`
The CMake Tools extension for VS Code that I use correctly, figured out
the developer presets when I include the file in my user presets.
However, running CMake in the command line did not work. It was
complaining about the dev-common preset not being reachable. As it turns
out, every preset file must directly include all preset files it
requires. It is not sufficient to include both the normal and the
developer presets in the user presets, because the developer presets
already require the normal presets.