Make `rf::Initialize()` return `Result<void>` again
The error handling for the initialization of the RF module is done in
the RF startup test thread and the SPI startup test and supervisor
thread. Therefore, `rf::Initialize()` must not use any kind of internal
error handling but just propagate any error up to the user.

I also fixed that the SPI supervisor thread only resetted the COBC when
there was a timeout error and increased the `startupTestTimeout`. The
latter was only 100 ms which is too short considering that the
`porCircuitSettleDelay` in `rf::Initialize()` alone is 100 ms.