Remove EDU communication error thread and rework error handling
I realized that the EDU communication error thread must not turn the EDU
on or off, because that is the responsibility of the power management
thread. Without this, however, the communication error thread is
obsolete since waiting till the EDU is alive is something all threads
communicating with the EDU must be able to do anyway. This is really
nice because it simplifies the code and most likely saves RAM and ROM.