Merge source files of convolutional coding
Previously, there was a source file for each of the two implementations:
the one with puncturing and the one without it. This lead to a lot of
code duplication and thereby a higher maintenance burden. For example,
when I added support for disabling the convolutional coding, I forgot
about the implementation without puncturing. Therefore, I merged both
source files into one now and use the preprocessor to enable the correct
code.