Fix `RepositoryContentSummaryReport::DoSize()`
`nObjects_` is the total number of objects in the repository which is
not necessarily the number of objects sent in a single report. Also
debug print an error message if adding a report to the Transfer Frame
fails. This should make it easier to find similar issues in the future.