Add min. RX timeout in RF communication thread
Previously, we only received additional data if we had time for the full
RX time of 3 s plus data handling. Now, we reduced this to a min. RX
time of 1 s plus data handling. The current implementation, however, is
too simple and does not compute the remaining RX time at all. Therefore,
a beacon can be sent up to 2 s too late. This is not too big of a deal,
though, especially since it will only happen if we want to receive "the
whole time".